Added interpolation submodule
Moved all interpolation routines to ~/dos/code/interpolation and added it as a
submodule for this repository. Needed a modification to burgersSLMM, currently
the only code which calls said routines.

Next would be to do the same for moving mesh codes. Hopefully this will still
allow me to pull over ssh to my laptop.