Added code to track speed during runs
Added a call to get_m_x in the time-step loop of burg2 and burgersSLMM and
output a vector with the position of the front.
This information is used by expt1 to get the line of best fit and hence use a
different method to find the speed of the front.
This information is then plotted by expt2.

~~This could be sped-up by removing the width-finding section of get_m_x.m,
which would remove two calls to fzero.~~ DONE

modified:   burg2.m
            burgersSLMM.m
            expt1.m
            expt2.m