Uses griddedInterpolant and removed big X_D bug
griddedInterpolant is used for linear, cubic, spline and pchip (spline
behaviour may be different). ENO currently not working.

For the limiter, created feval_lim.m, which is a bit simpler and more general
than the previously used interpolation/ppval_lim.m .

Bug: X_D previously used the value of U^n at the arrival point. Now corrected
to using the departure point. The interpolation is the same as that used for
RHS.