Fixed mistake in options files where BCs not used
Even though alpha_0 was being defined in the options files, it was having no
effect on the already defined u0, u_l and u_r.

Also changed the theta method to use theta_x = theta_t = 0.55, as it appeared
to not have any effect before - not tested if it changed.