Redone half the experiments with new changes
Experiments alan1, alan3, alan4 and alan5 complete.

Changes involve the new black and white colour scheme and the experiments
running for *all* N from 10 to 1000, instead of N from 20 to 1000, but
experiments getting sparser. The sparse experiments was a good idea to speed up
the run, since we're plotting it on a logarithmic scale.