Change feval_lim to use histc for extreme speedup
feval_lim now uses histc to find the data location, and a minmax to
limit. The has the effect that everything is vectorised and feval_lim
has gone from dominating the time for a limited burgersSLMM calculation
to being negligible.

It's also more terse, which maybe isn't a bad thing.